backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pydantic==1.10.12
motor==4.3.0
aiofiles==23.1.0
python-dotenv==1.0.0
Pillow==10.0.0
numpy>=1.24.0
pydicom>=2.4.0
opencv-python>=4.8.0
loguru==0.7.0
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.24.0
# MedGemmaService: Vertex AI / Hugging Face
google-cloud-aiplatform>=1.26.0
redis>=4.2.0
# Optional: for PHI redaction (pip install spacy && python -m spacy download en_core_web_sm)
# spacy
# en-core-web-sm
# Report generation (PDF/HTML)
Jinja2>=3.1.0
# WeasyPrint: requires system libs (Docker: libpango, libgdk-pixbuf, libcairo)
weasyprint>=60.0

# Optional (only if you will run MedGemma locally / HF):
# transformers==4.38.0
# torch==2.1.0
